# docs/generated/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BNPI PMS API",
    "version": "1.0.0",
    "description": "Welcome to the BNPI PMS API documentation. This REST API powers BNPI PMS backend services.",
    "contact": {
      "name": "BNPI PMS Support",
      "email": "support@bnpi-pms.example",
      "url": "https://bnpi-pms.example/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://bnpi-pms.example/terms"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    },
    {
      "url": "https://api.bnpi-pms.example/api",
      "description": "Production server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for authentication"
      }
    },
...
  },
  "paths": {
...
    "/api/sequential/{id}": {
      "get": {
        "summary": "Get sequential by ID",
        "description": "Retrieve a specific sequential by its unique identifier with optional field selection",
        "tags": [
          "Sequential"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "Sequential ID (MongoDB ObjectId format)",
            "example": "507f1f77bcf86cd799439011"
          },
          {
            "in": "query",
            "name": "fields",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of fields to include",
            "example": "id,name,code,pattern,current"
          }
        ],
        "responses": {
          "200": {
            "description": "Sequential retrieved successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Sequential not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "summary": "Update sequential",
        "description": "Update sequential data by ID (partial update)",
        "tags": [
          "Sequential"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "Sequential ID",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  },
                  "pattern": {
                    "type": "string"
                  },
                  "current": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sequential updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Sequential not found"
          },
          "409": {
            "description": "Sequential with this name or code already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete sequential",
        "description": "Soft delete an sequential by ID",
        "tags": [
          "Sequential"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "Sequential ID",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "200": {
            "description": "Sequential deleted successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Sequential not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sequential": {
      "get": {
        "summary": "Get all sequentials",
        "description": "Retrieve sequentials with advanced filtering, pagination, sorting, field selection, and optional grouping",
        "tags": [
          "Sequential"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of records per page"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Field to sort by",
            "example": "createdAt"
          },
          {
            "in": "query",
            "name": "fields",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of fields to include",
            "example": "id,name,code,pattern,current"
          },
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search query",
            "example": "INV"
          },
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            },
            "description": "JSON array of filter objects",
            "example": "[{\"code\":\"INV\"}]"
          },
          {
            "in": "query",
            "name": "groupBy",
            "schema": {
              "type": "string"
            },
            "description": "Group results by field",
            "example": "code"
          },
          {
            "in": "query",
            "name": "document",
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            },
            "description": "Include sequential documents"
          },
          {
            "in": "query",
            "name": "pagination",
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            },
            "description": "Include pagination metadata"
          },
          {
            "in": "query",
            "name": "count",
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            },
            "description": "Include total count"
          }
        ],
        "responses": {
          "200": {
            "description": "Sequentials retrieved successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create new sequential",
        "description": "Create a new sequential with the provided data",
        "tags": [
          "Sequential"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "code",
                  "pattern"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Invoice Number"
                  },
                  "code": {
                    "type": "string",
                    "example": "INV"
                  },
                  "pattern": {
                    "type": "string",
                    "example": "INV-{YYYY}-{0000}"
                  },
                  "current": {
                    "type": "integer",
                    "default": 0,
                    "example": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sequential created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Sequential with this name or code already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}